tCriando operacao de instaciacao

+++ src/main/java/com/cursoSpringUdemy/resources/CategoriaResource.java
@@ -20,7 +20,7 @@ public class CategoriaResource {
 
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    public ResponseEntity <?> find(@PathVariable Integer id){
+    public ResponseEntity <?> find(@PathVariable Long id){
         Categoria obj= service.buscar(id);
         return ResponseEntity.ok().body(obj);
         /*
